+++ packages/entity/src/components/fields/DruxtFieldDatetimeDefault.vue
@@ -4,7 +4,7 @@
     v-bind="wrapper.props"
   >
     <!-- Label: Above -->
-    <div v-if="$scopedSlots['label-above']">
+    <div v-if="$slots['label-above']">
       <slot name="label-above" />
     </div>