+++ packages/entity/src/components/fields/DruxtFieldResponsiveImage.vue
@@ -5,13 +5,13 @@
     v-bind="wrapper.props"
   >
     <!-- Label: Above -->
-    <div v-if="$scopedSlots['label-above']">
+    <div v-if="$slots['label-above']">
       <slot name="label-above" />
     </div>
 
     <!-- Label: Inline -->
     <slot
-      v-if="$scopedSlots['label-inline']"
+      v-if="$slots['label-inline']"
       name="label-inline"
     />