+++ packages/views/test/mixins/sorts.test.js
@@ -1,9 +1,8 @@
-import { createLocalVue, mount } from '@vue/test-utils'
+import { shallowMount } from '@vue/test-utils'
 
 import { DruxtViewsSortsMixin } from '../../src/mixins/sorts'
 
 // Setup local vue instance.
-const localVue = createLocalVue()
 
 const component = {
   name: 'DruxtViewsSortsTest',
@@ -11,8 +10,20 @@ const component = {
   render: () => ({})
 }
 
+
+
 test('DruxtViewsSortsMixin', async () => {
-  const wrapper = mount(component, { localVue })
+  const wrapper = shallowMount(component, {
+    global: {
+      plugins: [Vuex],
+
+      // OR:
+      mocks: {
+        $store: store,
+      }
+    }
+  })
+  //const wrapper = mount(component, { localVue })
   expect(wrapper.vm.value).toStrictEqual(undefined)
 
   wrapper.vm.model = 'title'