+++ packages/router/CHANGELOG.md
@@ -1,5 +1,19 @@
 # druxt-router
 
+## 0.24.2
+
+### Patch Changes
+
+- Updated dependencies [2ae1d6d]
+  - druxt@0.15.0
+
+## 0.24.1
+
+### Patch Changes
+
+- Updated dependencies [45bc0b9]
+  - druxt@0.14.0
+
 ## 0.24.0
 
 ### Minor Changes